% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_app}
\alias{phi_app}
\title{Soft BART matrix transformation}
\usage{
phi_app(X_stand, anc, tau)
}
\arguments{
\item{X_stand}{A matrix.}

\item{anc}{A tree matrix.}

\item{tau}{A double.}
}
\description{
Soft BART matrix transformation
}
