% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_app_r.R
\name{phi_app_r}
\alias{phi_app_r}
\title{Phi matrix calcualtions for SMOTR_BART}
\usage{
phi_app_r(X_stand, anc, tau)
}
\arguments{
\item{X_stand}{A matrix.}

\item{anc}{A tree matrix.}

\item{tau}{A double.' @return A mattrix}
}
\description{
Phi matrix calcualtions for SMOTR_BART
}
